+++ src/components/manifold-resource-deprovision/manifold-resource-deprovision.tsx
@@ -7,6 +7,7 @@ import { GetResourceQuery } from '../../types/graphql';
 
 @Component({ tag: 'manifold-resource-deprovision' })
 export class ManifoldResourceDeprovision {
+  @Prop() disabled?: boolean;
   @Prop() gqlData?: GetResourceQuery['resource'];
   @Prop() loading?: boolean = true;
 
@@ -17,6 +18,7 @@ export class ManifoldResourceDeprovision {
   render() {
     return (
       <manifold-data-deprovision-button
+        disabled={this.disabled}
         resourceId={this.gqlData && this.gqlData.id}
         resourceLabel={this.gqlData && this.gqlData.label}
         loading={this.loading}